refactor/헤더, 사이드메뉴 리팩토링
