+++ src/shared/Headers/components/LogoForHeader.tsx
@@ -1,0 +1,10 @@
+import Logo from '@/components/Logo';
+
+export default function LogoForHeader() {
+  return (
+    <>
+      <Logo size='sm' className='xl:hidden' />
+      <Logo size='lg' className='hidden xl:block' />
+    </>
+  );
+}
