+++ src/shared/Headers/components/UserIcon.tsx
