+++ src/shared/Headers/components/UserInfo.tsx
@@ -11,7 +11,7 @@ interface UserInfo {
 export default function UserInfo({ image, nickname, onClick }: UserInfo) {
   return (
     <div
-      className='flex cursor-pointer items-center justify-between gap-6'
+      className='flex justify-between items-center gap-6 cursor-pointer'
       onClick={onClick}
     >
       {image ? (
@@ -20,9 +20,9 @@ export default function UserInfo({ image, nickname, onClick }: UserInfo) {
           styles='w-24 h-24 border-1 border-solid border-gray-200 rounded-full'
         />
       ) : (
-        <IconUserSigned className='h-16 w-16 xl:h-24 xl:w-24' />
+        <IconUserSigned className='w-16 xl:w-24 h-16 xl:h-24' />
       )}
-      <p className='cursor-pointer font-[Pretendard] text-13 font-semibold leading-22 text-gray-300 xl:text-14 xl:leading-24'>
+      <p className='font-primary text-gray-300 text-13 xl:text-14 font-semibold leading-22 xl:leading-24 cursor-pointer'>
         {nickname}
       </p>
     </div>
